:zap::sparkles: feat(utils): add get_conf utility to get config Added get_conf method to get the logging configs from settings Renamed 'setup_conf' utility to 'setup_logging' Updated apps.py to apply get_conf method to set logging configurations
